warning removed
git-svn-id: https://qthreads.googlecode.com/svn/trunk@255 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ testincr.c
@@ -4,13 +4,13 @@
 
 aligned_t master = 0;
 
-aligned_t incr (qthread_t *me, void * arg)
+aligned_t incr(qthread_t * me, void *arg)
 {
     qthread_incr(&master, 1);
     return 0;
 }
 
-aligned_t incr5 (qthread_t *me, void* arg)
+aligned_t incr5(qthread_t * me, void *arg)
 {
     qthread_incr(&master, 5);
     return 0;
@@ -26,27 +26,27 @@ int main()
     qthread_incr(&master, 1);
     assert(master == 1);
     master = 0;
-    for (i=0; i<30; i++) {
+    for (i = 0; i < 30; i++) {
 	qthread_fork(incr, NULL, &(rets[i]));
     }
-    for (i=0;i<30;i++) {
-	qthread_readFF(NULL, NULL, rets+i);
+    for (i = 0; i < 30; i++) {
+	qthread_readFF(NULL, NULL, rets + i);
     }
     if (master != 30) {
-	printf("master is %lu rather than 30\n", master);
+	printf("master is %lu rather than 30\n", (long unsigned)master);
     }
-    assert (master == 30);
+    assert(master == 30);
     master = 0;
-    for (i=0; i<30; i++) {
+    for (i = 0; i < 30; i++) {
 	qthread_fork(incr5, NULL, &(rets[i]));
     }
-    for (i=0;i<30;i++) {
-	qthread_readFF(NULL, NULL, rets+i);
+    for (i = 0; i < 30; i++) {
+	qthread_readFF(NULL, NULL, rets + i);
     }
     if (master != 150) {
-	printf("master is %lu rather than 150\n", master);
+	printf("master is %lu rather than 150\n", (long unsigned)master);
     }
-    assert (master == 150);
+    assert(master == 150);
 
     qthread_finalize();
